Add changelog for version 0.9.1
This was cherry-picked from the 0.9.1 release, where I made the mistake
to add one commit for all updated stuff, instead of adding the changelog
in one commit and all other stuff in other commits.

But because we should have that information on master, we cherry-pick
the commit 7137a3ce7 partially here.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>
(cherry picked from commit 7137a3ce7)