Fix: Sort by date, not by id
This patch fixes a bug. We stated in the helptext of the "show" command
that when showing several logs, we sort by date. This was not true as we
sorted by Id, which also included the name of the log, hence we
effectively sorted by name.

This patch fixes that. Output is not sorted by date.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>